+++ optional_configuration/obituaries/obituaries_blufor.sqf
