+++ optional_configuration/obituaries/obituaries_civ.sqf
