+++ optional_configuration/obituaries/obituaries_global.sqf
