+++ optional_configuration/obituaries/obituaries_indfor.sqf
