+++ optional_configuration/obituaries/obituaries_opfor.sqf
